pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "glint"
version = "0.1.0"
description = "OPSEC digital footprint auditing framework"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0",
    "httpx>=0.27",
    "dnspython>=2.6",
    "python-dotenv>=1.0",
]

[project.scripts]
glint = "glint.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["glint*"]
